README.md: OSX users need to create an alias to localhost
OSX will not let a user connect to a locally running instance of samba. It says "This file server is available on your computer. Access the volumes and files locally."

So in order to let the user still connect they can create an alias to 127.0.0.1 and then it doesn't complain.